# Improve TMap's performance