--- a/benchmarks/src/main/scala/zio/stm/TMapContentionBenchmarks.scala
+++ b/benchmarks/src/main/scala/zio/stm/TMapContentionBenchmarks.scala
@@ -1,6 +1,5 @@
 package zio.stm
 
-import java.lang.{ Runtime => JRuntime }
 import java.util.concurrent.TimeUnit
 
 import org.openjdk.jmh.annotations._
@@ -16,23 +15,22 @@ import zio._
 class TMapContentionBenchmarks {
   import IOBenchmarks.unsafeRun
 
-  @Param(Array("0", "10", "100", "1000", "10000", "100000"))
-  var size: Int = _
+  @Param(Array("100", "1000", "10000"))
+  var repeatedUpdates: Int = _
 
   private var mapUpdates: UIO[Unit] = _
   private var refUpdates: UIO[Unit] = _
 
   @Setup(Level.Trial)
   def setup(): Unit = {
-    val data     = (1 to size).toList.zipWithIndex
-    val map      = unsafeRun(TMap.fromIterable(data).commit)
-    val ref      = ZTRef.unsafeMake(data.toMap)
-    val schedule = Schedule.recurs(1000)
-    val cpus     = JRuntime.getRuntime().availableProcessors()
-    val updates  = List.fill(cpus)(size - 1)
+    val schedule     = Schedule.recurs(repeatedUpdates)
+    val keysToUpdate = (1 to 100).toList
+    val data         = (1 to 1000).toList.zipWithIndex
+    val map          = unsafeRun(TMap.fromIterable(data).commit)
+    val ref          = ZTRef.unsafeMake(data.toMap)
 
-    mapUpdates = UIO.foreachParN_(cpus)(updates)(i => map.put(i, i).commit.repeat(schedule))
-    refUpdates = UIO.foreachParN_(cpus)(updates)(i => ref.update(_.updated(i, i)).commit.repeat(schedule))
+    mapUpdates = UIO.foreachPar_(keysToUpdate)(i => map.put(i, i).commit.repeat(schedule))
+    refUpdates = UIO.foreachPar_(keysToUpdate)(i => ref.update(_.updated(i, i)).commit.repeat(schedule))
   }
 
   @Benchmark